refactor(TurnkeyClient): split generated code and custom logic into +Extensions
